perf(masking): T50 crop to the AOI footprint before the where, bit-for-bit
mask_data_by_geometry applied a full-grid .where, so "mean NO2 over New
Jersey" against a continental grid reduced over an array that is >99% NaN by
construction. It now narrows the data -- and the already-rasterized mask --
to the mask's own footprint before the .where. All five call sites (stat,
plot, comparison, validation, export) inherit it with no call-site churn.
Measured 5.7x on a 9.6M-cell continental grid; identical numbers.

Rasterization deliberately stays on the FULL grid rather than moving behind
the crop as the PRD sketched. It costs ~11ms of a 100ms call and reads no
data, and keeping it there makes the kept cells identical to the uncropped
path's *by construction*. The PRD's premise -- that a cropped grid is a
subset with the same step, so its mask is the full-grid mask restricted --
does not survive float32 coordinate axes: re-rasterizing from cropped
endpoints derives a step differing in the 8th digit, which flips any cell
center lying on the geometry's boundary. Caught against the real TEMPO New
Jersey granule, where a Newark-sized AOI moved 2%; anchoring the affine to
the full grid did not fix it (GDAL's scanline arithmetic decides ties on the
window it is handed), cropping only the data does.

Cropping to the mask's footprint rather than the geometry's bbox also drops
the padding question entirely: T42's all_touched boundary cells are already
in the mask, so they are inside the window by definition.

area_weighted_mean now builds its cos(latitude) weights in float64. Real
granules publish float32 latitude, and float32 weights make the answer depend
on how many zero-weight masked-out cells are summed -- 5e-7 on a real TEMPO
New Jersey mean, which is exactly the crop-invariance this change needs.

Verified: 1095 Docker-green; crop=True/crop=False bit-identical (values,
count, min/max, percentile, area-weighted mean) on the delivered TEMPO
granule for both a geocoded NJ polygon and the knife-edge Newark box; live
end-to-end via the test account returns the same 3.315e+15 molecules/cm^2
and the same map, with aoi_crop_applied firing on both the statistic and the
plot.

Refs: docs/prds/prd-t50-crop-before-mask.md

+++ Backend/preprocessing/aggregation_service.py
@@ -181,7 +181,12 @@ def area_weighted_mean(da: xr.DataArray) -> float:
     if lat_name is None or lat_name not in da.dims:
         return float(np.nanmean(values[np.isfinite(values)]))
 
-    weights = np.cos(np.deg2rad(da[lat_name]))
+    # float64 weights regardless of how the granule stores its latitude axis:
+    # real products publish float32 lat, and accumulating float32 weights makes
+    # the answer depend on how many zero-weight (masked-out) cells happen to be
+    # summed -- the same region answered ~5e-7 differently from a continental
+    # granule than from a tight crop of it.
+    weights = np.cos(np.deg2rad(da[lat_name].astype("float64")))
     result = float(da.weighted(weights).mean(skipna=True))
     if not np.isfinite(result):
         raise ValueError("No finite values available for statistic.")
